Add files via upload
Data Cleaning Project Readme
This data cleaning project utilizes Python's Pandas library to enhance the quality and usability of a given dataset, focusing on columns such as age, rating, location, and salary. The primary objectives were accomplished through a series of structured steps:

The first task involved handling missing values in the 'age' and 'rating' columns. Missing age values were imputed with the mean age of the dataset, ensuring a representative and unbiased estimation. Similarly, missing rating values were replaced with the mean rating, contributing to a more complete and informative dataset.

The 'location' column underwent standardization to ensure consistency in formatting. This process aimed to resolve variations in how location data was recorded, providing a cleaner and more uniform representation.

A critical aspect of the cleaning process involved handling -1 values in a specific column. These values were replaced with boolean values, contributing to a clearer and more interpretable dataset.

Lastly, the 'salary' column, which originally contained a mix of currency symbols and 'k' values, was transformed into integer format. This transformation simplified the data and prepared it for further analysis, making it easier to understand and work with.